Extract loadout getting function
