+++ src/locale/en.json
@@ -617,7 +617,7 @@
       "BadURL": "Not a valid loadout share URL.",
       "Error": "Error getting loadout:",
       "Error404": "This loadout doesn't exist.",
-      "PasteHere": "Paste a loadout link to open the loadout in DIM."
+      "PasteHere": "Paste a loadout link to open the loadout."
     },
     "ImportLoadout": "Import Loadout",
     "ItemErrorSummary": "1 item error:",
